+++ react/components/EditorContainer/Sidebar/BlockEditor/hooks.ts
@@ -1,6 +1,7 @@
 import { equals, path } from 'ramda'
 import { useCallback } from 'react'
 import { defineMessages } from 'react-intl'
+import { useRuntime } from 'vtex.render-runtime'
 
 import { ConfigurationStatus } from '.'
 import {
@@ -19,6 +20,7 @@ import {
   omitUndefined,
   throttledUpdateExtensionFromForm,
 } from './utils'
+import { createEventObject } from '../../../../utils/auditEvents'
 
 const messages = defineMessages({
   cancel: {
@@ -61,6 +63,7 @@ export const useFormHandlers: UseFormHandlers = ({
   iframeRuntime,
   intl,
   saveContent,
+  sendEventToAudit,
   setState,
   showToast,
   state,
@@ -69,6 +72,7 @@ export const useFormHandlers: UseFormHandlers = ({
   const editor = useEditorContext()
   const formMeta = useFormMetaContext()
   const modal = useModalContext()
+  const { account, workspace } = useRuntime()
 
   const handleStatusChange = () => {
     const status = state.status ?? statusFromRuntime
@@ -200,6 +204,30 @@ export const useFormHandlers: UseFormHandlers = ({
 
       formMeta.setWasModified(false)
 
+      if (newConfiguration.status === ConfigurationStatus.SCHEDULED) {
+        const event = createEventObject(
+          'Schedule change',
+          'content',
+          account,
+          workspace,
+          contentId
+        )
+        await sendEventToAudit({
+          variables: { input: event },
+        })
+      }
+
+      const event = createEventObject(
+        'Edit content block',
+        'content',
+        account,
+        workspace,
+        contentId
+      )
+      await sendEventToAudit({
+        variables: { input: event },
+      })
+
       handleFormClose()
     } catch (err) {
       console.error(err)
@@ -230,6 +258,7 @@ export const useFormHandlers: UseFormHandlers = ({
     statusFromRuntime,
     iframeRuntime,
     saveContent,
+    sendEventToAudit,
     formMeta,
     handleFormClose,
     modal,
@@ -309,6 +338,17 @@ export const useFormHandlers: UseFormHandlers = ({
           template: editor.blockData.template,
           treePath: editor.blockData.serverTreePath,
         },
+      })
+
+      const event = createEventObject(
+        'Activate content block version',
+        'content',
+        account,
+        workspace,
+        blockId
+      )
+      await sendEventToAudit({
+        variables: { input: event },
       })
     } catch (err) {
       console.error(err)