+++ react/components/EditorContainer/Topbar/ContextSelectors/BindingCloning/BindingCloningModal.tsx
@@ -9,6 +9,7 @@ import {
 } from 'vtex.styleguide'
 import { defineMessages, FormattedMessage, useIntl } from 'react-intl'
 import { pick } from 'ramda'
+import { useRuntime } from 'vtex.render-runtime'
 
 import BindingSelector, { BindingSelectorItem } from './BindingSelector'
 import {
@@ -17,6 +18,10 @@ import {
   useCloneContent,
 } from './CloneContentContext'
 import OverwriteDialog, { useOverwriteDialogState } from './OverwriteDialog'
+import { createEventObject } from '../../../../../utils/auditEvents'
+import SendEventToAuditMutation, {
+  SendEventToAuditMutationFn,
+} from '../../../mutations/SendEventToAudit'
 
 interface Props {
   isOpen?: boolean
@@ -104,6 +109,7 @@ const BindingCloningModal: FunctionComponent<Props> = ({ isOpen, onClose }) => {
     saveRoute,
     copyBindings,
   } = actions
+  const { account, workspace } = useRuntime()
 
   const checkOverwrites = () => {
     return new Promise<void>((resolve, reject) => {
@@ -218,8 +224,10 @@ const BindingCloningModal: FunctionComponent<Props> = ({ isOpen, onClose }) => {
 
   const handleConfirmCopyContent = async ({
     showToast,
+    sendEventToAudit,
   }: {
     showToast: ShowToastFunction
+    sendEventToAudit: SendEventToAuditMutationFn
   }) => {
     if (submitStatus === 'SUBMITTING') {
       return
@@ -237,6 +245,17 @@ const BindingCloningModal: FunctionComponent<Props> = ({ isOpen, onClose }) => {
       })
       try {
         await applyChanges()
+
+        const event = createEventObject(
+          'Copy binding content',
+          'binding',
+          account,
+          workspace
+        )
+        await sendEventToAudit({
+          variables: { input: event },
+        })
+
         setTimeout(() => {
           showToast(intl.formatMessage(toastMessages.success))
         }, 500)
@@ -269,12 +288,21 @@ const BindingCloningModal: FunctionComponent<Props> = ({ isOpen, onClose }) => {
                     <FormattedMessage id="admin/pages.editor.components.button.cancel" />
                   </Button>
                 </span>
-                <Button
-                  disabled={submitStatus === 'SUBMITTING'}
-                  onClick={() => handleConfirmCopyContent({ showToast })}
-                >
-                  <FormattedMessage id="admin/pages.editor.components.button.save" />
-                </Button>
+                <SendEventToAuditMutation>
+                  {sendEventToAudit => (
+                    <Button
+                      disabled={submitStatus === 'SUBMITTING'}
+                      onClick={() =>
+                        handleConfirmCopyContent({
+                          showToast,
+                          sendEventToAudit,
+                        })
+                      }
+                    >
+                      <FormattedMessage id="admin/pages.editor.components.button.save" />
+                    </Button>
+                  )}
+                </SendEventToAuditMutation>
               </div>
             }
           >